Connect frontend and backend
