+++ backend/index.js
@@ -1,9 +1,22 @@
 const express = require("express");
+const axios = require("axios");
 
 const app = express();
 
-app.get("/api/v1/quest", (req, res) => {
-  res.send("working");
+const url = "/api/v1/quest";
+const answer = [];
+
+app.get(url, (req, res) => {
+  res.send(answer);
+});
+
+app.post(url, (req, res) => {
+  answer.push(req.body);
+
+  res.status(200).json({
+    status: "success",
+    data: answer,
+  });
 });
 
 const port = 5001;
