--- frontend/README.md
@@ -1,1 +1,0 @@
-# flavour-quest
