+++ frontend/package.json
@@ -13,6 +13,7 @@
     "@chakra-ui/react": "^2.7.0",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
+    "axios": "^1.4.0",
     "chakra-react-select": "^4.6.0",
     "formik": "^2.4.2",
     "framer-motion": "^10.12.16",
