--- frontend/src/pages/Results.tsx
@@ -1,3 +1,0 @@
-export function Results() {
-  return <> xxxx </>;
-}
