+++ frontend/vite.config.ts
@@ -6,5 +6,8 @@ export default defineConfig({
   plugins: [react()],
   server: {
     port: 3000,
+    proxy: {
+      "/api": "http://localhost:5001",
+    },
   },
 });
